adicionado classes de mapeamento banco de dados

+++ Versoes.Core.Domain/Entities/Usuario.cs
@@ -2,12 +2,12 @@
 {
     public class Usuario
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string Nome { get; set; }
         public StatusDeCadastro Status { get; set; }
         public string Login { get; set; }
         public string Senha { get; set; }
-        public long SetorId { get; set; }
+        public int SetorId { get; set; }
         public Setor Setor { get; set; }
     }
 }
